requirements.txt
azure-identity
azure-search-documents
azure-storage-blob

nltk

openai
langchain_community

python-dotenv

flask
flask-cors
